Change: remove scanconfig library
As the transformation from xml to json is just done within the nasl-cli
it doesn't make sense to express this functionality.
