--- rust/scanconfig/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "scanconfig"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-storage = { path = "../storage" }
-models = { path = "../models" }
-serde = { version = "1.0", features = ["derive"]}
-quick-xml = { version = "0.28.1", features = ["serialize"] }
-serde_json = "1.0"
-urlencoding = "2.1.2"
-tracing = "0.1.37"
